se modificaron archivos
Se modificaron archivos

+++ php/eliminar.php
@@ -1,9 +1,10 @@
 <?php
     require("conexion.php");
-
-   $id_edificio = $_POST["id_edificio"];
-   
-    $delete = "DELETE FROM edificio WHERE id_edificio='$id_edificio'";
-    $result = mysqli_query($conexion,$delete);
-
+    if (isset($_POST['id_edificio'])){
+        $id_edificio = $_POST["id_edificio"];
+        $delete = "DELETE FROM edificio WHERE id_edificio='$id_edificio'";
+        $result = mysqli_query($conexion,$delete);
+    }else{
+        header("Location: login.php");
+    }
 ?>
